Redo Aave linear factory v2 deployment task (#2124)
* Rename task, delete artifacts

* Update build info

* Fix task code

* Add changelog entry

* Update artifacts

* Update pkg/deployments/CHANGELOG.md
