+++ pkg/deployments/action-ids/arbitrum/action-ids.json
@@ -437,28 +437,6 @@
       }
     }
   },
-  "20221115-aave-rebalanced-linear-pool": {
-    "AaveLinearPoolFactory": {
-      "useAdaptor": false,
-      "actionIds": {
-        "create(string,string,address,address,uint256,uint256,address)": "0xce1c9e518c0270b5db510d1cd33336a3d6a6d80154e5535dcf4825442712d291",
-        "disable()": "0x7d9e561eb57049f7aa87abe6ccd3a88c062d0112ae53f060cf1aeac13d17a567"
-      }
-    },
-    "AaveLinearPool": {
-      "useAdaptor": false,
-      "factoryOutput": "0x77f0e64Faf0FDeAF0b632919214ba142f0478424",
-      "actionIds": {
-        "disableRecoveryMode()": "0x15f00bddcd311bbb6cb92c87b273ded5f162a286cbc5fdd8f20f6a9e6e1c9fff",
-        "enableRecoveryMode()": "0xed5630bca392a4970a1901079b4d679f7a22b124ac0fbf88e17c8a3a4daa267a",
-        "initialize()": "0xd3457c3b7bf9fb1f545d7d8c25d7b088ab15e0a71a3c5359d4b9c3827c0e19d6",
-        "pause()": "0x718fffbe2c8b4df95b0f67dd732370c9e580255923b4ecc837aafe975e17403e",
-        "setSwapFeePercentage(uint256)": "0xf12947894def448d0c9e33d6220a4f23cbbdb269739ac99695817db884648f5e",
-        "setTargets(uint256,uint256)": "0xbe17db3a209eabd6173aeb955c3e5d02b6c71f7453406425e1b1ec733233081a",
-        "unpause()": "0x26d530e934979106bf47db262f8f2ec443528a326fc167183dd5895f16fac8e8"
-      }
-    }
-  },
   "20221124-authorizer-adaptor-entrypoint": {
     "AuthorizerAdaptorEntrypoint": {
       "useAdaptor": false,
