+++ pkg/deployments/action-ids/mainnet/action-ids.json
@@ -913,28 +913,6 @@
       }
     }
   },
-  "20221115-aave-rebalanced-linear-pool": {
-    "AaveLinearPoolFactory": {
-      "useAdaptor": false,
-      "actionIds": {
-        "create(string,string,address,address,uint256,uint256,address)": "0x5f96988dd5367f81278bc9c31c9abb4549e8de284234707f9e86994c45c292ab",
-        "disable()": "0x3924d0d790727bf2925421c7e316cfbe3d8b69f26b36b9d7d1c97e32bdeb4947"
-      }
-    },
-    "AaveLinearPool": {
-      "useAdaptor": false,
-      "factoryOutput": "0xe23D111cb71414830B93F078977B7f53A5e996eE",
-      "actionIds": {
-        "disableRecoveryMode()": "0xd3eb886ad36bb38e7b7109377afb7b290db25f057b6f8f4bc098df1996e614ab",
-        "enableRecoveryMode()": "0x6a688dda46d2337436e2ffa772e0c7510eb072d5c45225400a19a61bd76360e4",
-        "initialize()": "0x81b3fafbd77a3e501466178dfd836697ab4ee6d272d8e05a2f7b1759d1318bc1",
-        "pause()": "0x1f16abe3860c7a3426659e50f0217af96ac40aa554d8ddaebcb7c399118eeb1b",
-        "setSwapFeePercentage(uint256)": "0x0693774dcda5e82a5b5f4255fe8bc7aa5f7ce39cd6b4f9986b116fc4af317450",
-        "setTargets(uint256,uint256)": "0x881bd2702150eafb9524fe01e983df0fb0e99eca758c1b3959e46a084cc1618b",
-        "unpause()": "0xcdd7ab46c8258e8c091144b92a3a1061315e0da3aef7773d859de4ee421fd920"
-      }
-    }
-  },
   "20221124-authorizer-adaptor-entrypoint": {
     "AuthorizerAdaptorEntrypoint": {
       "useAdaptor": false,
