+++ pkg/deployments/action-ids/polygon/action-ids.json
@@ -522,28 +522,6 @@
       }
     }
   },
-  "20221115-aave-rebalanced-linear-pool": {
-    "AaveLinearPoolFactory": {
-      "useAdaptor": false,
-      "actionIds": {
-        "create(string,string,address,address,uint256,uint256,address)": "0xa1d068db796ebd013ec3aa0c761d682aa19f6eb0964b6c9ba1a99ad8009a3b2e",
-        "disable()": "0xa1a1644c63123b10ec06491be8dfc9942647bc3e7ee28e80b69eafc36bdea79b"
-      }
-    },
-    "AaveLinearPool": {
-      "useAdaptor": false,
-      "factoryOutput": "0xfCCcB77A946b6a3BD59d149F083B5BfbB8004D6D",
-      "actionIds": {
-        "disableRecoveryMode()": "0xe4b2d6d888f0825b65c76622fb2ecc6db98ca5bae83250c67ff1fb8b4af5340e",
-        "enableRecoveryMode()": "0xc66ec5b1a0165981818e33d19a4b5613006a1b432746a54ccefd9d7b5c4e63b3",
-        "initialize()": "0x667b0bf7b294b45986fb92cd4bd9cdd21aaad496bc49f888b231c50e780788c0",
-        "pause()": "0xd9f2ebf54cfbfc103a3e56c91eea301de16b390949bab52a9651e8e69ea43338",
-        "setSwapFeePercentage(uint256)": "0xeb10286296a5305f41da3ebe2f35b24016d6b8ea9afea03d53e74cc6a9f595ac",
-        "setTargets(uint256,uint256)": "0x96f17bf5369c2ad1494d20a0c2a169ed5757a75b7416eae507024b364525f7b5",
-        "unpause()": "0xe526c58cabb37c617299105b1267f5ef8ed12edbf391b6f80c6f7d655e732bc6"
-      }
-    }
-  },
   "20221124-authorizer-adaptor-entrypoint": {
     "AuthorizerAdaptorEntrypoint": {
       "useAdaptor": false,
