--- pkg/deployments/tasks/20221115-aave-rebalanced-linear-pool/output/goerli.json
@@ -1,6 +1,0 @@
-{
-  "AaveLinearPoolFactory": "0x4C4287b07d293E361281bCeEe8715c8CDeB64E34",
-  "MockAaveLendingPool": "0x6d471c05f8A99Bd8f991467A8283e3DC968b8D7c",
-  "MockStaticAToken": "0x5c89d9E1A668d0B39F4EC242F280bdDd445C633c",
-  "MockAaveLinearPool": "0x88D07558470484c03d3bb44c3ECc36CAfCF43253"
-}
