+++ pkg/deployments/tasks/deprecated/20220817-aave-rebalanced-linear-pool/readme.md
@@ -2,7 +2,7 @@
 
 > ⚠️ **DEPRECATED: do not use** ⚠️
 >
-> This relayer has been deprecated in favor of an [updated version](../../20221115-aave-rebalanced-linear-pool), which improves the Rebalancer contract to handle more tokens and fixes a potential issue where a malicious contract being called while fetching the wrapped token rate could extract value from the pool.
+> This relayer has been deprecated in favor of an [updated version](../../20221207-aave-rebalanced-linear-pool-v3), which improves the Rebalancer contract to handle more tokens and fixes a potential issue where a malicious contract being called while fetching the wrapped token rate could extract value from the pool.
 
 Deployment of the `AaveLinearPoolFactory`, for Linear Pools with a wrapped aToken. This new deployment includes a permissionless rebalancing contract that has Asset Manager privileges.
